Make GoGet and GoInstall useable from generators
goGet/goInstall are unexported in the cmd package, which makes then unusable for commands on the generate package (ie: goth generator).

This PR adds GoGet and GoInstall to the generate package and updates existing generators to use the new exported version.

This PR also moves the previous goGet/goInstall to root.go (to avoid a package namespace conflict with the soda generator) and changes them to simple call genetate.Go*.