Add post initializers support
PostInitializers introduces a new way to handle a common pattern:
when you want to init some stuff or start a service just after the app
started successfully, you usually just drop the relevant code after the
routes init, in the `if app == nil {` block.

This way to do is just a workaround, and doing so can result in multiple
init, if you use `App()` more than once.

PostInitializers proposes a cleaner solution for this pattern. It allows
the user to define a slice of handlers to execute after the app starts.