Fix an issue when using --with-dep but dep is not there
Due to the way exec.Command works, the command lookup was done before
the generator execution, so even if we have a step to go get dep before
running the goGet() step, it will always fallback to go get instead of
using dep.

This issue is only visible once, since it needs to have dep missing, which
is installed during the first run.