Fail if github username can't be found
When we execute the git command to see if we can get the github's username we
need to check that the err is nil and not that the err is not nil as we were
doing now. This was causing that the command was failing in case that
github.username couldn't be fetched causing username to always be empty and the
help message to be wrong because the template was missing "username".

Example: ... "$GOPATH/src/github.com//myfabulousproject" (adjust accordingly).