Skip -i flag in go build if modules are enabled
Using -i flag when running go build is not useful with go modules enabled
and in some cases can lead to a bug, with go trying to rebuild
shipped vendorized dependencies as described in this issue
golang/go#27285.
With this commit we check if go modules are enabled and we drop the -i
flag accordingly.