plugins: move addToCache to avoid bad entries
This came from an issue referenced on slack (gophers #buffalo on 2019-07-12) where I could not get buffalo plugins, especially pop, to load properly.

It happened because my machine timed out trying to run `buffalo-pop available`, and it cached a bad entry with no commands.

It seems like we should only cache the commands if we got a successful response from the given binary.

The downside here is that if there is a `buffalo-<something>` binary and we're failing to get commands from it, buffalo will try again on it every time. I've added logging for that error so it's clear to the user why buffalo is loading slowly and so they can fix the issue.