Fix generated GitLab ci YAML (#2086)
* gitlab-ci: Default to Golang 1.15

Given the policy by the golang team to support and maintain only the
latest two Go releases defaulting to Go 1.8 effectily made buffalo test
with a EOL'ed Golang.

This change defaults to 1.15, and allows the user to switch to 1.16 with
the already present latest switch.

* gitlab-ci: Leverage gomodules over GOPATH

Prior to this change, the `.gitlab-ci.yml` linked the directories so
that GOPATH would work. Given buffalo now always uses gomodules these
steps are no longer required, and currently even break the pipeline
created from the generated CI definitions.

By just executing the job in the `/builds` subdirectory gomodules work
as expected again.

* gitlab-ci: Update postgres connection details

Docker links the containers differently than was priorly assumed for
GitLab CI. As such the out-of-the-box settings weren't properly working.
This change resolves that in the CI YAML definitions to isolate the
changes from local setups.