For nulls.Time, decode empty value as NULL
This commit adds a check during decoding which will return `nulls.Time` as
a NULL value (i.e. its value for `Valid` will be false). Without this
check, the value of nulls.Time{} will be:
```
{
  Time: 0001-01-01 00:00:00 +0000 UTC,
  Valid: true,
}
```
This ends up storing the value `0001-01-01 00:00:00 +0000 UTC` in the
database, as opposed to a NULL value.